% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cols.R
\name{check_cols}
\alias{check_cols}
\title{Check checker dataframe column-wise for proper number of suppressed values.}
\usage{
check_cols(checker, supp_col)
}
\arguments{
\item{checker}{Dataframe to be analyzed, product of checker_df().}

\item{supp_col}{Columns to be checked.}
}
\value{
An integer vector.
}
\description{
Check checker dataframe column-wise for proper number of suppressed values.
}
\examples{
x <-
 data.frame(
   x = c(1, 0, 0),
   y = c(0, 1, 1),
   z = c(0, 1, 0)
)

check_cols(x, c('x', 'y', 'z'))
check_cols(x, c('z'))
}
