% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_df.R
\name{checker_df}
\alias{checker_df}
\title{Check a dataframe for current suppression}
\usage{
checker_df(df, supp_cols, regex_suppchar)
}
\arguments{
\item{df}{Dataframe to be checked.}

\item{supp_cols}{Columns to be checked for suppression.}

\item{regex_suppchar}{Suppression character to check for (special char. processed)}
}
\value{
A dataframe.
}
\description{
Check a dataframe for current suppression
}
\examples{
x <- data.frame(x = c('2', '*', '*', '3', '5', '2'), y = c('*', '3', '5', '*', '2', '*'))
supp_col <- c('x','y')
regex_suppchar <- '\\\\*'
}
