% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementary.R
\name{complementary}
\alias{complementary}
\title{Recursively fix insufficiently-suppressed rows and columns in dataframe}
\usage{
complementary(df, supp_col, supp_col_idx, regex_char)
}
\arguments{
\item{df}{Dataframe to be checked and fixed for necessary suppression.}

\item{supp_col}{Vector of columns to be checked for suppression.}

\item{supp_col_idx}{Indices of supp_col in df.}

\item{regex_char}{Character to replace suppressed value.}
}
\value{
A dataframe.
}
\description{
Recursively fix insufficiently-suppressed rows and columns in dataframe
}
\examples{
x <-
 data.frame(
   x = c('-', '6', '8'),
   y = c('-', '6', '-'),
   z = c('10', '-', '9')
)

complementary(x, c('x', 'y', 'z'), c(1,2,3), '-')
}
