% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_row.R
\name{fix_row}
\alias{fix_row}
\title{Fix row with insufficient number of suppressed values.}
\usage{
fix_row(df, i, supp_col_idx, rc_char)
}
\arguments{
\item{df}{Dataframe with row to be fixed.}

\item{i}{Index of row to be fixed.}

\item{supp_col_idx}{Vector of indices of suppression columns.}

\item{rc_char}{Non-special character to replace suppressed value.}
}
\value{
A dataframe.
}
\description{
Fix row with insufficient number of suppressed values.
}
\examples{
x <-
 data.frame(
   x = c('-', '6', '8'),
   y = c('-', '6', '-'),
   z = c('10', '0', '9')
)

fix_row(x, 3, c('x', 'y', 'z'), '-')
fix_row(x, 3, c('x', 'y') , '-')
}
