% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_col_idx.R
\name{supp_col_idx}
\alias{supp_col_idx}
\title{Return indices of dataframe suppression columns.}
\usage{
supp_col_idx(df, supp_col)
}
\arguments{
\item{df}{Dataframe to be scanned.}

\item{supp_col}{Columns to be searched for.}
}
\value{
An integer vector.
}
\description{
Return indices of dataframe suppression columns.
}
\examples{
x <-
 data.frame(
   x = c(1, 2, 3),
   y = c(3, 4, 5),
   z = c(5, 6, 7)
)

supp_col_idx(x, c('x', 'y'))
supp_col_idx(x, c('z'))
}
