% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress.R
\name{suppress}
\alias{suppress}
\title{Use complementary suppression to protect small cell counts.}
\usage{
suppress(df, supp_val, supp_char, supp_col, totals = "none")
}
\arguments{
\item{df}{Dataframe to be fixed for necessary suppression.}

\item{supp_val}{Maximum value to be suppressed (all values equal to or smaller suppressed).}

\item{supp_char}{Character to be used to indicate suppressed value.}

\item{supp_col}{Vector of columns to be checked for suppression.}

\item{totals}{"rows", "totals", c("rows", "totals"), or "none" Will dictate which totals are returned with the df.}
}
\value{
A dataframe.
}
\description{
Use complementary suppression to protect small cell counts.
}
\examples{
x <-
 data.frame(
   x = c(1, 6, 7),
   y = c(2, 6, 0),
   z = c(10, 4, 9)
)

suppress(x, 5, '-', c('x', 'y', 'z'), "row")
}
